refactor hooks directory
