+++ protocol-designer/src/components/organisms/TipPositionModal/hooks/usePositionReference.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useSelector } from 'react-redux'
 import { round } from 'lodash'
 
 import { DropdownMenu } from '@opentrons/components'
@@ -10,19 +9,9 @@ import {
   POSITION_REFERENCE_TOP,
 } from '@opentrons/shared-data'
 
-import { getRobotType } from '../../../file-data/selectors'
-import {
-  getAdditionalEquipmentEntities,
-  getLabwareEntities,
-  getPipetteEntities,
-} from '../../../step-forms/selectors'
-import { getLiquidClassesValues } from '../../../steplist/formLevel/handleFormChange/utils'
-
 import type { Dispatch, SetStateAction } from 'react'
 import type { DropdownOption } from '@opentrons/components'
-import type { PositionReference, WellLocation } from '@opentrons/shared-data'
-import type { FormData } from '../../../form-types'
-import type { MoveLiquidPrefixType } from '../../../resources/types'
+import type { PositionReference } from '@opentrons/shared-data'
 
 interface UsePositionReferenceResult {
   positionReferenceDropdown: JSX.Element
@@ -121,35 +110,5 @@ export function usePositionReference(args: {
     ),
     reference,
     setReference,
-  }
-}
-
-export function useDefaultPosition(
-  formData: FormData | null,
-  prefix: MoveLiquidPrefixType
-): WellLocation {
-  const pipetteEntities = useSelector(getPipetteEntities)
-  const labwareEntities = useSelector(getLabwareEntities)
-  const additionalEquipmentEntities = useSelector(
-    getAdditionalEquipmentEntities
-  )
-  const robotType = useSelector(getRobotType)
-  if (formData == null) {
-    return {}
-  }
-  const liquidClassDefaultValues = getLiquidClassesValues({
-    rawForm: formData,
-    pipetteEntities,
-    labwareEntities,
-    additionalEquipmentEntities,
-    robotType,
-  })
-  return {
-    origin: liquidClassDefaultValues[`${prefix}_position_reference`],
-    offset: {
-      x: liquidClassDefaultValues[`${prefix}_x_position`],
-      y: liquidClassDefaultValues[`${prefix}_y_position`],
-      z: liquidClassDefaultValues[`${prefix}_mmFromBottom`],
-    },
   }
 }
